lib, zebra, doc: extend ibgp support to nhg in zebra
When routes from eBGP peers are pushed in ZEBRA, and are not
resolved over connected paths, an error message is displayed.
This is because neither the ALLOW_RECURSION, nor the IBGP flag
are set on the route. When attempting to do the same in SHARP,
it appears that the NHG do not have this IBGP flag.

> cli# debug zebra rib detail
> [..]
> 2023/12/17 21:09:43 ZEBRA: [NWWQT-S584X]         nexthop_active: Route Type bgp has not turned on recursion 172.31.0.102/32 failed to match
> 2023/12/17 21:09:43 ZEBRA: [PKF9S-7G8XM]         EBGP: see "disable-ebgp-connected-route-check" or "disable-connected-check"

Attempting to push NHGs to ZEBRA does not lead to the same
error messages, for two reasons:
- NHGs are not sent by SHARPD if allow-recursion flag is not set
- the IBGP flag is not implemented in NHGs

This commit includes two changes:
- introduce the ZEBRA_NHG_FLAG in the nexthop_group structure
The flag will be used by ZEBRA when attempting to check if
the nexthops of the NHG are active or not.
- add a vty command in nexthop group that maps over the IBGP
flag.

> ubuntu2204(config)# nexthop-group H
> ubuntu2204(config-nh-group)# ibgp

The below test shows that when turning on ibgp, the EBGP trace
disappears:

> ubuntu2204(config)# debu zebra nexthop de
> ubuntu2204(config)# debu zebra rib de
> ubuntu2204(config)# ip route 172.31.0.102/32 192.168.0.77
> ubuntu2204(config)# nexthop-group H
> ubuntu2204(config-nh-group)# force-nexthop-config
> ubuntu2204(config-nh-group)# nexthop 172.31.0.102
> [..]
> 2023/12/17 22:09:43 ZEBRA: [NWWQT-S584X]         nexthop_active: Route Type sharp has not turned on recursion 172.31.0.102/32 failed to match
> 2023/12/17 22:09:43 ZEBRA: [PKF9S-7G8XM]         EBGP: see "disable-ebgp-connected-route-check" or "disable-connected-check"
> ubuntu2204(config-nh-group)# ibgp
> 2023/12/17 22:14:57 ZEBRA: [NWWQT-S584X]         nexthop_active: Route Type sharp has not turned on recursion 172.31.0.102/32 failed to match

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>