bgpd: add nhg group support for multiple path routes
The nexthop group support is not applicable to ECMP BGP paths.
For instance, when BGP addpath is configured, only prefixes
associated to a nexthop are using nexthop groups at BGP level.

Unlock the possibility to use nexthop groups for ECMP paths.
The choice is done to create a nexthop group hierarchy with
multiple nexthops.

For each path to install, two nexthop groups will be used:
nexthops and groups.
- Link the parent nexthop group to the path (with pi->bgp_nhg)
- Link the original bgp_nhg_cache nexthop to the path (with pi->bgp_nhg_nexthop).
- Handle the installation order at proto level: the nexthop first, the parent after
- Add nexthop extension in 'show bgp nexthop-group'
- unlock the ability to handle multiple paths with NHGs
- A double hash list is used to handle frr_each_safe() limitations:
this technical choice is done in order to be able to detach both
nexthops and groups in a single walk. Otherwise, asan errors are
detected.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>